requirements.txt
bcrypt==4.3.0
cffi==1.17.1
coverage==7.8.0
cryptography==44.0.2
greenlet==3.2.1
iniconfig==2.1.0
packaging==25.0
paho-mqtt==2.1.0
pluggy==1.5.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
PyQt6==6.9.0
PyQt6-Qt6==6.9.0
PyQt6_sip==13.10.0
pyserial==3.5
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-cov==6.1.1
python-dotenv==1.1.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
